Update Bootstrap to alpha 6
Updates Bootstrap to alpha 6 and fixes issues with navigation bar.
The problem was that in alpha 6 navbar needs the `navbar-toggleable-md`
class. And the logout link needs to be in a separate `ul` in order to be
floated to the right because now the navbar uses flexbox.

Adds a hamburger icon for the collapsed navigation.