Fix category show page issues
Adds truncation for long item names.

Buttons have same length so they can be grouped.
Buttons are positioned relative to the bottom of the card so that the
image size doesn't move them around.

Buttons next to the variation name are always on the same line. Either
next to the variation name or on the next line.

A little margin is added between the rows of cards for better visual
effect.

Card dimensions are tweaked a little bit so that they look better on
both mobile and desktop.

Card image size is changed a little bit so that longer images won't go
outside of the card box.